chore(weave): Extract common functions in `eval.py` into re-usable common modules. (#3328)
* initial refactor

* Larger Refactor

* Comments!
